remove workspace context handler
