+++ vscode/microsoft-kiota/src/commands/regenerate/regenerateButtonCommand.ts
@@ -4,8 +4,8 @@ import { ExtensionContext } from "vscode";
 import { extensionId, treeViewId } from "../../constants";
 import { getExtensionSettings } from "../../extensionSettings";
 import { getGenerationConfiguration, setGenerationConfiguration } from "../../handlers/configurationHandler";
-import { WorkspaceGenerationContext } from "../../handlers/workspaceGenerationContextHandler";
 import { OpenApiTreeProvider } from "../../openApiTreeProvider";
+import { WorkspaceGenerationContext } from "../../types/WorkspaceGenerationContext";
 import { isClientType, isPluginType } from "../../util";
 import { confirmOverride } from "../../utilities/regeneration";
 import { Command } from "../Command";
