+++ vscode/microsoft-kiota/src/types/WorkspaceGenerationContext.ts
@@ -1,0 +1,7 @@
+import { ClientOrPluginProperties } from "../kiotaInterop";
+
+export interface WorkspaceGenerationContext {
+  clientOrPluginKey: string;
+  clientOrPluginObject: ClientOrPluginProperties;
+  generationType: string;
+}
